refactor the test_verify_blob_checksum test
Use the checksums actually attached to the source files.  In the case of test_bad_source_data/incorrect_checksum, the checksums are intentionally set incorrectly so that the test will flag that as failures.

This should address #156.
