+++ tests/test_gcshcablobstore.py
@@ -8,14 +8,13 @@ import unittest
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..')) # noqa
 sys.path.insert(0, pkg_root) # noqa
 
-from dss.blobstore import BlobNotFoundError
 from dss.blobstore.gcs import GCSBlobStore
-from dss.hcablobstore import HCABlobStore
 from dss.hcablobstore.gcs import GCSHCABlobStore
 from tests import utils
+from tests.test_hcablobstore import HCABlobStoreTests
 
 
-class TestGCSHCABlobStore(unittest.TestCase):
+class TestGCSHCABlobStore(unittest.TestCase, HCABlobStoreTests):
     def setUp(self):
         self.credentials = utils.get_env("GOOGLE_APPLICATION_CREDENTIALS")
         self.test_bucket = utils.get_env("DSS_GCS_TEST_BUCKET")
@@ -25,33 +24,6 @@ class TestGCSHCABlobStore(unittest.TestCase):
 
     def tearDown(self):
         pass
-
-    def test_verify_blob_checksum(self):
-        self.assertTrue(
-            self.hcahandle.verify_blob_checksum(
-                self.test_src_data_bucket, "test_good_source_data/0",
-                {
-                    HCABlobStore.MANDATORY_METADATA['CRC32C']['keyname']: "e16e07b9",
-                }
-            )
-        )
-
-        self.assertFalse(
-            self.hcahandle.verify_blob_checksum(
-                self.test_src_data_bucket, "test_good_source_data/1",
-                {
-                    HCABlobStore.MANDATORY_METADATA['CRC32C']['keyname']: "e16e07b9",
-                }
-            )
-        )
-
-        with self.assertRaises(BlobNotFoundError):
-            self.hcahandle.verify_blob_checksum(
-                self.test_src_data_bucket, "DOES_NOT_EXIST",
-                {
-                    HCABlobStore.MANDATORY_METADATA['CRC32C']['keyname']: "e16e07b9",
-                }
-            )
 
 
 if __name__ == '__main__':
